+++ src/components/TypeComponents/SimpleSearch/Form/formParsing/widgetToInitialValues.js
@@ -19,7 +19,7 @@ const widgetToInitialValues = (widget) => {
             });
           }
           // At this point, we have a token, so we need to parse it out to work out the other fields
-          const dateMatch = tokenMatch.match(/(currentDate)((#)(-?)(\d{1,3}))?((#)([d,w,m,y]))?/);
+          const dateMatch = tokenMatch.match(/(currentDate)((#)(-?)(\d{1,3}))?((#)([d,w,M,y]))?/);
           return ({
             comparator: fcr.comparator,
             relativeOrAbsolute: 'relative',