+++ src/components/TypeComponents/SimpleSearch/Form/results/SimpleSearchResults.js
@@ -16,10 +16,10 @@ import {
   TextField
 } from '@folio/stripes/components';
 
-import RowWithDelete from '../../../WidgetComponents/misc/RowWithDelete';
+import RowWithDelete from '../../../../RowWithDelete';
 import SimpleSearchResultField from './SimpleSearchResultField';
-import SimpleSearchSort from '../sort/SimpleSearchSort';
-import DragAndDropFieldArray from '../../../DragAndDropFieldArray';
+import SimpleSearchSort from '../sort';
+import DragAndDropFieldArray from '../../../../DragAndDropFieldArray';
 
 import css from './SimpleSearchResults.css';
 
@@ -34,8 +34,6 @@ const SimpleSearchResults = ({
   id
 }) => {
   const intl = useIntl();
-
-
   const renderResultField = (fieldName, index, _droppable, _draggable, fields) => {
     return (
       <div className={css.resultLine}>