+++ src/components/Widget/WidgetFooter.test.js
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event';
 
 import '@folio/stripes-erm-components/test/jest/__mock__';
 import { renderWithIntl } from '@folio/stripes-erm-components/test/jest/helpers';
-import translationsProperties from '../../../../test/helpers';
+import translationsProperties from '../../../test/helpers';
 
 import WidgetFooter from './WidgetFooter';