+++ src/tokens.js
@@ -3,7 +3,7 @@
 // TOKENS can take the form
 // {{currentUser}} or
 // {{currentDate}} {{currentDate#15}} {{currentDate#-10}} {{currentDate#3#w}}
-// So currentDate, optionally +- up to 3 digits, then optionally a 'd', 'w', 'm' or 'y'
+// So currentDate, optionally +- up to 3 digits, then optionally a 'd', 'w', 'M' or 'y'
 
 import moment from 'moment';
 
@@ -24,20 +24,22 @@ function tokens(valueString, stripes, options = {}) {
   }
 
   // MATCH CURRENT DATE +- DAYS
-  const dateMatch = tokenMatch.match(/(currentDate)((#)(-?\d{1,3}))?((#)([d,w,m,y]))?/);
+  const dateMatch = tokenMatch.match(/(currentDate)((#)(-?\d{1,3}))?((#)([d,w,M,y]))?/);
   if (dateMatch?.[1] === 'currentDate') {
     // We have matched the date pattern, do date logic
     const currentDate = moment(new Date()).startOf('day');
+
     if (dateMatch?.[4]) {
+      const offsetInteger = parseInt(dateMatch[4], 10);
+
       // We have an integer to add to date
       if (dateMatch?.[7]) {
-        // Add days, months or years
-        return currentDate.add(dateMatch?.[4], dateMatch?.[7]).format(dateFormat);
+        // Add days, months, weeks or years
+        return currentDate.add(offsetInteger, dateMatch[7]).format(dateFormat);
       }
       // Assume days if not
-      return currentDate.add(dateMatch?.[4], 'd').format(dateFormat);
+      return currentDate.add(offsetInteger, 'd').format(dateFormat);
     }
-
     return currentDate.format(dateFormat);
   }